🔧 Commit local changes in `run` commands
Although these will be replaced by new top-level commands, we want to commit all local
changes before removing the `run` commands, so we don't lose any code that we might
have not transferred.

We'll skip linting for this commit. No need to fix linting issues for code we'll remove
after.