✨ Add the `CifCleanWorkChain` controller
Add a first version of the `CifCleanWorkChain` controller, used to run the second
import from the MPDS.

The controller hard-codes a lot of inputs for now. The user can only specify the codes
for the COD tools and the groups to put the parsed `StructureData` and clean `CifData`
in. This was mainly to set up the submissions quickly, without taking too much time to
build a very flexible API. That said, it might make sense to hard code the inputs for
this part of the infrastructure, since we want to be consistent across databases etc.