✨ Add "source"-related tools
As is to be expected, we work a lot with "sources" in the `mc3d-source` package. The
"source" is typically a dictionary with three keys:

* `database`: The source database (`mpds`, `icsd`, ...)
* `version`: The version of the database ID at import.
* `id`: The source ID, also extracted from the database.

Since it's easier to work with strings e.g. when using sets, we often convert the
source into a "source string", here written as an f-string with the keys above:

f'{database}|{version}|{id}', e.g. 'MPDS|1.2.0|S123456'

This commit adds a whole bunch of tools in the form of functions:

* `get_source_uuid_mapping`
* `get_source_string`
* `get_spacegroup_number`
* `find_source_structure`
* `find_source_structure_cached`
* `get_source_structure_dict`
* `find_cif_clean`
* `sources_match`

See the corresponding docstrings for more details.