✨ Add "deprecation" analysis commands
As we keep on updating the database, there might be several reasons to deprecate a
source:

* Corresponding ID is no longer in the source database.
* We notice an issue that makes the structure likely invalid (e.g. incorrect formula).
* ...

Here we add CLI tools for analysing these issues, and writing them to an output file
(default: `deprecation.json`), which can be updated by the other commands. This way
we can build a single file that stores all deprecated sources and why they are
deprecated.

To formalise the different reasons for deprecating a source, we add an `Enum`:
`SourceDeprecation`. We also add an `DEPRECATION_HELP` that explains in more detail
what each mode of deprecation means. This is put in a `constants.py` module at the
root, which can also host other variables which we want to keep consistent throughout
the package.