✨ Add `select` command
Once we have obtained a list of unique families of duplicates, we need to select the
structures from the new families we have run. The new command takes care of this.

Here we add an initial draft of the command. Since it is still possible that the logic
might change, we leave the documentation for once its stabilizes.