📦 Add missing dependencies
The package dependencies were woefully inadequate, since it was mostly used in a project
environment that had everything installed. Here we add the missing dependencies, save
for `scipy`. Only a single component of the codebase relies on it, so we temporarily
catch the `ImportError` and indicate that the package should be installed, until we
figure out which methods we want to keep.